authorization-service/fast-api/app/requirements.txt
motor==3.0.0
tornado>6.0.0, <7.0.0
python-dotenv==0.20.0